RENAME album property to playlist, minor refactor of layout
