+++ __tests__/PlayInfo/Comments.spec.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { shallow } from 'enzyme';
-import Comments from '../../client/src/components/Comments/Comments';
+import Comments from '../../client/src/components/PlayInfo/Comments';
 
 xdescribe('Test Suite', () => {
   it('should render without throwing an error', () => {
