+++ sampleDataGen/dataGen.js
@@ -100,7 +100,10 @@ for (let i = 0; i < trackCount; i += 1) {
 
   track.id = i;
   track.submittedDate = faker.date.past(3);
-  track.album = randName() + randAppend();
+  track.playlist = {
+    name: randName() + randAppend(),
+    art: `https://loremflickr.com/640/480/abstract?fake=${Math.floor(Math.random() * 100)}`,
+  };
   track.albumArt = `https://loremflickr.com/640/480/abstract?fake=${Math.floor(Math.random() * 100)}`;
   track.title = randName() + randAppend();
   track.artist = randName(); // aka artist
